/simplify: one dialRecorder for the /solid seam tests, and an honest provenance line
Four-angle cleanup pass on the url-default commit: the dial-capture harness
the new test copied from readout.test.ts is one shared
dialRecorder.testlib.ts now (both files import it); the tautological second
assertion (re-deriving the expected URL with the same helper the
implementation calls) is dropped in favour of the literal; and the option
docstring + Reference line no longer claim "every browser consumer of this
seam" spelled the derivation — the in-tree browser consumers dial
connectSurfaces, which deliberately keeps its required url; the hand-spelling
this default deletes lives in the downstream reference consumer.
