+++ website/src/content/surface/ref-surface-app.mdx
@@ -288,10 +288,10 @@ tag for the watchdog to probe.
 
 `connectSurface`'s `url` is **optional**: omitted, it defaults to
 `surfaceWsUrl(location.origin)` — the page's own origin through the one
-scheme-swap + path derivation, which is what every browser consumer spelled by
-hand (a browser app dials the origin that served it; the derivation was never a
-choice). Omitting it with no `location` in scope (a Node caller) throws loudly —
-pass the URL you actually mean there.
+scheme-swap + path derivation, the value a browser consumer spells by hand
+otherwise (never a choice: a browser app dials the origin that served it).
+Omitting it with no `location` in scope (a Node caller) throws loudly — pass
+the URL you actually mean there.
 
 Tags that ride the same wire but are **not** sibling surfaces go in
 `connectSurfaces`' `extraGroups` — a keyed map's group (for
